requirements.txt
absl-py
annotated-types
antlr4-python3-runtime
attrs
certifi
charset-normalizer
chex
click
cloudpickle
contourpy
cv
cycler
decorator
distrax
dm-tree
docker-pycreds
etils
farama-notifications
filelock
flax
fonttools
fsspec
gast
gitdb
gitpython
gym
gym-notices
gymnasium
gymnax
huggingface-hub
humanize
hydra-core
idna
imageio
importlib-resources
iniconfig
kiwisolver
markdown-it-py
matplotlib
mdurl
ml-dtypes
msgpack
nest-asyncio
numpy
omegaconf
opencv-python
opt-einsum
optax
orbax-checkpoint
packaging
pandas
pillow
platformdirs
pluggy
protobuf
psutil
pydantic
pydantic-core
pygments
pyparsing
pytest
python-dateutil
pytz
pyyaml
regex
requests
rich
safetensors
scipy
seaborn
sentry-sdk
setproctitle
setuptools
simplejson
six
smmap
submitit
tensorboardx
tensorflow-probability
tensorstore
tokenizers
toolz
tqdm
transformers
typing-extensions
tzdata
urllib3
wandb
zipp
scikit-learn
pqdm
pipe